+++ opentelemetry-api/src/opentelemetry/context/__init__.py
@@ -28,7 +28,7 @@ _available_contexts = {}
 
 for entry_point in iter_entry_points("opentelemetry_context"):
     try:
-        _available_contexts[entry_point.name] = entry_point.load()
+        _available_contexts[entry_point.name] = entry_point.load()  # type: ignore
     except Exception as err:  # pylint: disable=broad-except
         logger.warning(
             "Could not load entry_point %s:%s", entry_point.name, err
