+++ opentelemetry-sdk/setup.py
@@ -44,7 +44,7 @@ setuptools.setup(
     include_package_data=True,
     long_description=open("README.rst").read(),
     long_description_content_type="text/x-rst",
-    install_requires=["opentelemetry-api==0.4.dev0", "aiocontextvars"],
+    install_requires=["opentelemetry-api==0.4.dev0"],
     extras_require={},
     license="Apache-2.0",
     package_dir={"": "src"},
